+++ Source/UnrealSharpScriptGenerator/Program.cs
@@ -17,7 +17,6 @@ public static class Program
 	public static string EngineGluePath => ScriptGeneratorUtilities.TryGetPluginDefine("GENERATED_GLUE_PATH");
 	public static string PluginDirectory => ScriptGeneratorUtilities.TryGetPluginDefine("PLUGIN_PATH");
 	public static string ManagedBinariesPath => Path.Combine(PluginDirectory, "Binaries", "Managed");
-	public static string ManagedPath => Path.Combine(PluginDirectory, "Managed");
 	public static string ScriptFolder
 	{
 		get
@@ -59,8 +58,8 @@ public static class Program
 			// Everything is exported. Now we need to compile the generated C# code, if necessary.
 			if (FileExporter.HasModifiedEngineGlue)
 			{
-				Console.WriteLine("Compiling generated engine C# glue...");
-				PublishSolution(Path.Combine(ManagedPath, "UnrealSharp"));
+				string engineGluePath = Path.Combine(EngineGluePath, "UnrealSharp");
+				DotNetUtilities.BuildSolution(engineGluePath);
 			}
         
 			stopwatch.Stop();
@@ -74,63 +73,6 @@ public static class Program
 			Console.WriteLine("Stack Trace:");
 			Console.WriteLine(ex.StackTrace);
 			Console.ResetColor();
-		}
-	}
-	static string FindDotNetExecutable()
-	{
-		var pathVariable = Environment.GetEnvironmentVariable("PATH");
-    
-		if (pathVariable == null)
-		{
-			throw new Exception("Couldn't find dotnet.exe!");
-		}
-    
-		var paths = pathVariable.Split(Path.PathSeparator);
-    
-		foreach (var path in paths)
-		{
-			// This is a hack to avoid using the dotnet.exe from the Unreal Engine installation directory.
-			// Can't use the dotnet.exe from the Unreal Engine installation directory because it's .NET 6.0
-			if (!path.Contains(@"\dotnet\"))
-			{
-				continue;
-			}
-			
-			var dotnetExePath = Path.Combine(path, "dotnet.exe");
-			
-			if (File.Exists(dotnetExePath))
-			{
-				return dotnetExePath;
-			}
-		}
-
-		throw new Exception("Couldn't find dotnet.exe!");
-	}
-	
-	static void PublishSolution(string projectRootDirectory)
-	{
-		if (!Directory.Exists(projectRootDirectory))
-		{
-			throw new Exception($"Couldn't find project root directory: {projectRootDirectory}");
-		}
-		
-		string dotnetPath = FindDotNetExecutable();
-		
-		Process process = new Process();
-		process.StartInfo.FileName = dotnetPath;
-		
-		process.StartInfo.ArgumentList.Add("publish");
-		process.StartInfo.ArgumentList.Add($"\"{projectRootDirectory}\"");
-		
-		process.StartInfo.ArgumentList.Add("-warn:1");
-		process.StartInfo.ArgumentList.Add($"-p:PublishDir=\"{ManagedBinariesPath}\"");
-		
-		process.Start();
-		process.WaitForExit();
-		
-		if (process.ExitCode != 0)
-		{
-			throw new Exception($"Failed to publish solution: {projectRootDirectory}");
 		}
 	}
 }
